Add check for program validity before using it to paint
Add check for program validity before using it to paint. It's
possible that the program has been created and is available as part
of the device state but is in an invalid state (i.e the program has
failed to build/link). This will then cause GL error when trying to
draw with this program.