Disable sync to vblank and use timer to throttle gfx widget
Disable sync to vblank and use timer to throttle gfx widget. When
sync to vblank is enabled it creates a performance problem when
we have multiple windows open and each is rendered in turn and
each render ends with a vblank synchronized swap. This halves the
frame rate for each new window !

This problem is inherent to using a single thread and to render
to multiple window surfaces.

The current solution is to disable the vblank sync (allows swap
to return immediately) and then in order to avoid rendering at some
rediculous fps (my Geforce 1650 renders at ~10k fps) and burning
the CPU/GPU we need to throttle the rendering by using a timer.