Add confirmation of closing a mainwidget with unsaved changes
Add confirmation of closing a mainwidget with unsaved changes. For
each of the resource type (material, particle engine, animation) we
can compute a hash value based on what the resource contained before
and then compare to the current hash value before proceeding to
close the tab and possibly ask the user for confirmation whether
he wants to save or not.

Refactor the overwrite checking to check if the resource being saved
is the "same" as what is currently being edited. I.e. use a hash
value to differentiate between case where user wants to modify some
value in resource X from the case when user creates some new resource
Y that accidentally has the same name as X.