Define the distinction between primitive and not primitive resource
Define the distinction between primitive and not primitive resource.
A primitive resource is a resource that exists by default and isn't
created by the user. For example Checkerboard texture material or
solid color material or Rectangle shape.

Rename the Workspace methods to list the materials and drawables
and provide new methods to list only primitive materials and drawables.