Add workspace (project) settings and the settings dialog
Add workspace (project) settings and the settings dialog. The idea
that these settings will/can be used in the following cases:

- During the rendering of the content in the editor. Applicable
  settings are MSAA and texture min/mag filter

- When launching the game from the editor in "play mode" (TB).
  All settings would apply.

- When packaging the workspace content the settings could/would be
  written in the exported JSON and the game launcher would then
  perform the window and rendering context creation using these
  settings/parameters.

This commit only adds the settings object and the serialization
to/from JSON and the dialog for modifying the settings. The actual
use of the data as outlined above needs to be implemented in other
subsequent work/commits.