Refactor rendering timer into mainwindow from gfx widget
Refactor rendering timer into mainwindow from gfx widget. When the
"mainloop" of triggering the rendering is in the mainwindow it's
possible to support logic that needs to be done once per rendering
loop iteration easily. For example if/when using a single device
and device context the GPU resource garbage collection should be
done only once per iteration (or at some steady interval) instead
of every GfxWidget invoking the same GC method.