Refactor game state loading and saving away from main function
Refactor game state loading and saving away from main function.

This is a step towards a design where there is an external application
that launches and runs the game and sets up the appropriate resources
for the game/application to consume. The game state loading is specific
to the game and is thus better to be located into the game specific
code which removes these game specific details from the main function.