Add first implementation of main game/app runner in gamelib
Add first implementation of main game/app runner in gamelib. The
runner (also know as a "host") is given a JSON file, which specifies
the initial configuration such as the OpenGL configuration and
window properties. The application will then try to create the rendering
context with the config, the window and the rendering device as per
the configuration data.

It will the proceed to load a shared library, which is expected to
contain an entry point for bootstrapping the actual game/application
creation. The entry point (MakeApp) will be invoked in order to
create a new App instance. Once this instance exists the host
application will enter the main application loop and start performing
the following activities:

- Process any pending platform specific window messages.
- Process any incoming requests from the application.
- Track the passing of wall and game/application time.
- Drive the frame rendering and simulation updates by calling
  the app implementations for Update/Tick and Render.
- Do some simple book keeping about rendering statistics.

The app instance is also asked for a wdk::WindowListener instance
which (when given) is connected to the window events. This is the
primary mechanism the application can respond to windowing events
coming from the platform. In addition to this the application is
obviously free to use whatever other means it wants to in order
to for example poll the keyboard state.

There are two functions that are invoked for the purposes of updating
the game state.  Update and Tick.

- Update is expected to be the "high frequency" function that updates
  the game simulations and animations. It's expected to be running
  at a resolution closer to the rendering rate. The update loop should
  take less real wall time to run than what is the current update time
  step. I.e. if the Update runs at 30HZ then the time step is 1.0/30.0
  ~0.0333...s. If the Update takes longer than that things will run
  into trouble.

- Tick is expected to be the "low frequency" function that ticks low
  frequency functions such as AI simulations or such. It is expected
  to be running at low resolution such as 1HZ. However even with Tick
  the same notes about the wall time vs. time step apply as with Update.

The current design of keeping time is as follows.

Displaying a frame takes some amount of time which we don't really care
about. This time is measured and used to take any number of fixed delta
Update steps before the next frame is to be rendered. I.e. the rendering
time of the previous frame is used to drive the updates for the *next*
frame.  The "leftover" time is accumulated until there's enough to
take yet another simulation update step. This should in theory provide
decent performance in case the game cannot keep up the frame rates and
is rendering bound. For example assume that a vsync is missed then the
app will miss the signal and the frame won't show up until at the next
vblank. This will half the rendering/frame rate but the as long as the
simulation can keep up simulating X amount of time in less than X real
time things should keep working. If the application is CPU/simulation
bound things will likely fail. In the opposite (more likely case) that
the rendering workload is too small and there's no VSYNC we might actually
need to add some arbitrary wait/limiters in order to throttle the
execution and save some energy/CPU/GPU cycles.

During execution the application *may* request the host application
for some services related to for example window management. These can
be expressed by pushing them into a queue which the host application
will inspect on every iteration of the main loop. These are indeed
*requests* and the application should not assume any particular outcome.

The current plan is that there will be an alternative host/main
implementation that will be in part of the Gamestudio Editor. That will
use a Qt provided window and set things up so that the game can access
resources in the current workspace.

TODO: Currently there's no provision for explicitly enabling/disabling
sync to vblank. This feature would need to be implemented in wdk/

TODO: Currently there's no way to query the current display refresh rate.
This could be useful for cases when we want to for example run Update
at display refresh frequency.

TODO: Throttle the main loop in case where vblank is turned off and
running "too fast" with plenty of time to spare.

References:
https://gafferongames.com/post/fix_your_timestep/
Game Engine Architecture by Jason Gregory