Refactor global logger access into a LockedLogger class
Refactor global logger access into a LockedLogger class. Depending
on the build setup there might be multiple binaries being produced
using the same logger.cpp translation unit which means there'll be
multiple copies of the same global objects. Therefore having a global
mutex to protect the access to the logger might no longer actually
work since there could be multiple mutexes.  If the locking code
is moved into the object that we pass around a pointer to then this
problem can go away. (Other solution would be to change the way
binaries are being built and have the common code refer to a shared
library that would contain the shared functionality).