Add first implementation for playing the game/app in the editor
Add first implementation for playing the game/app in the editor.

Previous commits have made changes in the way the games/apps are
launched, i.e. the application logic can be built into a shared
library. We can leverage this structure and also perform load the
library in the Editor application and then setup a new window and
rendering surface and start calling methods on the game::App
implementation. The goal of this is to allow quicker and easier
iteration of the game when building the game content in the Editor.
I.e. one does not need to build/package the project and then go
about lauching the game separately but the game can be started
directly from the editor and ideally it should be able to display
the latest changes in the content as they're begin done in the
editor.

The first implementation of this "play app" functionality is in
PlayWindow, which is created in the MainWindow when play application
is requested.

The PlayWindow implementation will try to load the application library
as per the current project settings in the workspace. If that is
succesful a new game::App instance is created through a call to
MakeApp (which is a global function exported by the library). The
arguments set in the workspace project settings are then parsed
and passed to the game instance for processing and so on.
Overall the functionality is very similar to the game/application
runner that is in gamelib.

There are some current shortcomings, not yet implemented things and
bugs.

- The application isn't yet able to access the content directly from
  the Editor but it's still going to have to do the file based load.
  This needs to be changed so that the data can be transferred or made
  visible for the app without having to write/load anything from a file.

  It could be possible to provide the app with some kind of loader
  instance  (similar to that of gamelib loader class) but that would
  then force some dependencies to be built into game lib loader (or
  built into a yet another shared library)

  Alternative could be to have a mechanism where the content is serialized
  into JSON and the JSON is then passed to the app and the app
  creates the content objecs based on this in memory JSON instead
  of loading the JSON from the file.

- Working folder maddness. The app might expect to load data at runtime
  from a location reletive to the application's working folder.
  When the editor runs it has a different working folder. The current
  "solution" is to change the working folder temporarily before
  calling into the app. This is ugly and will certainly lead into
  some problems.

- Unresolved paths and working folder. For example the graphics subsystem
  uses paths (for shaders) that are without any resource encoding.
  If the current working folder is changed to the app's desired working
  folder then application data might be loaded OK but also the shader
  paths would only be valid if the app was packaged at least once
  or had the shaders copied by some other way.

  This issue is interconnected to the working folder issue in
  general.

- Process separation. Current issue is that the if the application
  faults in the library it'll bring down the editor too. This isn't
  nice. Therefore the PlayWindow should be built into a separate
  executable that gets launched from the Editor and which then
  goes on to load the app library. Getting the data from the
  Editor will then require some form of IPC.

- Mouse movement and other not yet supported window callbacks.
  The PlayWindow should hide the mouse pointer and pass the
  mouse events to the application.
  There's also a bunch of other window events that need to be
  forwarded.

- Window focus is incorrect from the start. Calling
  QWindow::requestActivate causes a X11 error to be printed in
  the console.