Share workspace content with the game/app when running in Editor
Share workspace content with the game/app when running in Editor.

This commit improves the ability to play the game/app in the Editor
by make it possible for the app to access the content directly from
the current workspace instead of loading it (again) from the file.

The above is accomplished by adding an interface for looking for
game content/assets and then providing the app instance an object
that implements this interface. Subsequently we can then have
alternative implementations, one that source the content from a
JSON file (the normal content loader implementation) and another
implementation which sources the content from the currently opened
workspace.

In order to simplify the current build configuration and to disable
the propagation of "unwanted" dependencies into the main runner
application there's a "stub" gamelib.cpp which contains a helper
function to create a default environment for the game content.
This stub functionality can then be built into the game/app shared
library which then keeps all the dependencies in this single library.
This however is just a current implementation detail and might change
at some point.