Split material state into dynamic and static state
Split material state into dynamic and static state and add static
flag. When the static flag is set to true the material uses íts
parameter values to derive a material ID which is then used to map
to the actual shader program. Then the static state (that is not
expected to change) is only applied and set *once* into the shader
program when the program is initially created.

This change reduces the calls we must make for setting glUniforms
in the shader program by turning those parameters into their own
shader program. This means that there will be more shder programs
and more GPU memory consumption however the performance is improved
when less state needs to be set in the shader program.

Changing any of the material parameters of the static material will
trigger the recompilation and rebuild of a new shader program. This
should not be done very frequently or performance will be degraded.
In fact if dynamic changes are required then it'd be better to not
mark the material as "static" in the first place.

This change also enables the shader program garbage collection since
we might now have excess of shader programs (if the materials are
changed) that are no longer needed or used. Those can be then reclaimed.