Enable sync to vblank (swap interval) on a single rendering surface
Enable sync to vblank (swap interval) on a single rendering surface.

When rendering to multiple windows there's a problem wrt how to manage
the rendering rate. If VSYNC is turned off the main rendering loop
(see main.cpp) will run as fast as possible and the application will
likely be rendering more frames than the display can actually show.
This leads to waste in terms of CPU and GPU processing. For example
when running on the laptop within 5 mins the fan is at 100% the
laptop is burning hot and the battery is drained. This is not good.
However if all the rendering surfaces have VSYNC setting enabled then
swapBuffers call will block on every window swap which means there
are as many waits (per second) as there are windows. This means that
every additional window will decrease the rendering rate.
For example let's say the display runs @ 60 Hz, then,
  2 windows -> two swaps -> 30fps
  3 windows -> three swaps -> 20 fps
  6 windows -> 6 swaps -> 10 fps.

Using timers or inserting waits anywhere in the rendering path will
lead to very janky rendering, so these cannot be used.

Ultimately it seems that the solution is to enable sync to VBLANK for
only a *single* rendering surface. Thus making sure that we're doing
only a single wait per second in the call to swap buffers. So what we
try to do here is to have a flag to tell us when some surface has vsync
enabled and if that surface is destroyed some other surface needs to
enable this flag.

Notes about Qt.
setSwapInterval is a member of QSurfaceFormat. QSurface and QOpenGLContext
both take QSurfaceFormat but for the swap interval setting only the data
set in the QSurface matters.
The implementation tries to set the swap interval on every call to the
platform's (GLX, WGL, EGL) "make current" call. However in the QWindow
implementation the QSurfaceFormat is only ever set to "requestedFormat"
which is only used *before* the native platform window is ever created.
This means that a subsequent call to QWindow::setFormat will not change
the swap interval setting as such but the native resources must be
destroyed and then re-created. (doh)

See: qwindow.cpp, qglxintegration.cpp

Some old discussion about the problem.
https://stackoverflow.com/questions/29617370/multiple-opengl-contexts-multiple-windows-multithreading-and-vsync

see commit. e14c0325af44512740ad711d0c13fd276efc25b7