Add primitive line drawable and high level line drawing methods
Add primitive line drawable and high level line drawing methods.
The line primitive defines two vertices that rasterize a horizontal
line. In order to draw a line between two arbitrary points A and B
one needs to work out the distance and angle between the two points
in order to transform the line drawable properly. This functionality
is implemented in the drawing.cpp (DrawLine).

Note that the line primitive isn't added to the Editor (for example
in the Animation widget) since it doesn't seem very useful of it
itself.