Add support for creating/storing/loading user defined shapes
Add support for creating/storing/loading user defined shapes.
User defined shapes use the Polygon drawable with triangle fans that
the user can create using the polygon widget. It's possible to use
a material as a "blueprint" i.e. to create a polygonal shape that
correspods to the shape in the material, i.e. the non-transparent
pixels of some sprite. Create a custom shape will then reduce the
overdraw when the shape is rendered as opposed to using some other
shape such as rectangle which will invoke rasterization for many
fragments that will end up sampling transparent pixels and thus have
no affect on the render color buffer.