Remove isExposed check in Editor's GfxWidget's paintGL method
Remove isExposed check in Editor's GfxWidget's paintGL method.
The problem with the isExposed check is that currently a sync to
vblank is required in order to throttle the main loop. If the gfx
widget rendering is skipped (including swap i.e. sync) then when
the window is not exposed and none of the application's windows with
gfx widgets are exposed then the application will no longer sync
to vblank, which in turn means that the editor's main loop will
turn into a busy loop and burn a lot of CPU cycles.

It looks like really the easist fix for this is to *not* check for
isExposed. However there have been some occasional qt log messages
about swap buffers being undefined on an unexposed window. These
happen at least in the dialogs such as DlgText that use a timer to
run the animation (they're not rendered as part of the main widget
rendering) and it's possible that the timer event to trigger the
repaint would happen *before* the widget (dialog) is actually show.
The solution here is to move the render timer start in the callback
(onInitScene) that is invoked from the GfxWidget when things are
ready for rendering.