Share Editor's MainWindow's workspace menu with the child windows
Share Editor's MainWindow's workspace menu with the child windows
so that the same functionality and same keyboard shortcuts work for
example for opening workspace resources or creating new resources.

In the MainWindow implementation instead of using "this" as the
parent for the dialogs the parent needs to be the currently active
window in case the action was invoked from a child window. Using
Mainwindow instance would set the mainwindow as the active window
before opening the dialog.