Refactor running an iteration of the game loop in the Editor
Refactor running an iteration of the game loop in the Editor. The
problem before is that calling the function to update the state
and then render directly from the main loop in main.cpp doesn't work
when a modal dialog is open. It doesn't work because a  modal dialog
will enter a temporary event loop while the dialog is being shown
and that means that the thread will not get a chance to update/render
the widgets.

This refactoring uses the same logic for the application main loop
but instead of triggering the game loop iteration directly the action
is queued to the application's main event queue. With this small change
the application can then run the update/render even when a modal
dialog is open.