Refactor resource types into type and instance classes
Refactor resource types into type and instance classes. Whenever
we want to create a new resource such as material what we really
want to create is a new type of a material. I.e let's say that the
new material type would be called "flappy bird" and would render a
sprite based animation. The material type would then contain the
properties such as the sprite set (the textures) that are relevant
to that type. However there could be several *instances* of that
material type each with their own per instance data (such as the
current animation runtime). Using a single class to represent both
the material type and the instance type is confusing and leads to
confusing code in cases such as copying (are we copying instance
representations or class reprensetations?) and also wastes memory
since the type data doesn't change between instances.

This change splits the resource classes into two classes. For
example Material class is now split into two classes.

MaterialClass
-------------
Instances of this class are used to represent new material types.
The expectation is that there is a single MaterialClass object
for each new material type in the material system. Instances of
this material type then share the class object pointer and can
refer to it for the static behaviour. (static not as per C++ type
but as per the material type). Furthermore it's expected that the
material class objects are created through the editor application
and managemed centrally by the applications asset "library", i.e.
they can be serialized and loaded to/from JSON.

Material
------------
Instances of this class are used to represent material instances
of some material type. Instances of the same material type all
share the same material class object and refer to the class object
for the static behaviour. Eeach instance however can contain data
that is specific to the instance only. For example with the example
material "flappy bird" each instance could have different current
time i.e. be at different stages in their animation cycle. The
time values can change indepedencly (per instance) whereas the
material's texture data is same for each instance (material type
data).

The same change has been done to all the following types:
Material, Drawable and Animation. During refactoring of the the
Animation class the previous functionality involving timelines
was removed since that needs to be re-thought anyway. In fact
new classes are also included for creating new "animation tracks"
with support for "actuators" that apply some change/transformation
on some animation node instance during the playback of the track.
Not all Drawable implementations yet have a class type since they
don't have any class/type specific data.

This change removes some functionality from the AnimationWidget
i.e. it's currently no longer possible to "play" the animation
in the widget as such. Instead the play functionality needs to
also be refactored to support playing animation tracks (or a
"null" track).

Also some other smaller refactorings include:

- Refactor the workspace live handle tracking. Using a shared_ptr
  to the resource class object inside the GameResource template
  allows this class object to be directly used for creating
  instances of resources.

Some things that probably should be refactored at some point.
- Packing of drawable types (does the shader packing make any
  sense?)
- The drawable style option. Generic implementations for wireframe
  and outline would be nice.