Remove universal event filter for zoom in/out functionality
Remove universal event filter for zoom in/out functionality. This
was added before to capture mouse scroll (with Control key) events
and turn them into zoom in/out requests universally. But this has
the problem that it's not possible to have several different widgets
each support this action without capturing and turning them all
into ZoomIn/ZoomOut calls.

For example the animation track widget has a timeline widget that
supports zoom in/out on mouse croll (with Control key). But on this
zoom in/out we don't want to zoom in/out on the main content that
is in the gfx widget. Therefore it's better to simply remove the
universal event filter and let the widgets capture this themselves.
Top level zoom in/out is then only triggered through the action
in the application menu.