Remove call to ConfirmClose on ChildWindow shutdown
Remove call to ConfirmClose on ChildWindow shutdown. This change
masks an issue (a crash) that happens / can happen because of
getting butt fucked by unwanted recursion. I.e. if the ChildWindow
is being closed and the closeEvent handler calls ConfirmClose which
pops a message box Qt will enter a message loop for the duration
of the QMessageBox. This message loop will then be processing the
event queue including main window's IterateGameLoopEvent which when
processed will call back into the child window (and widget).

By not calling ConfirmClose this particular instance of this type
of bug can be resolved. ConfirmClose should not be called at
application shutdown anyway since the state of all the currently
open widgets is preserved. ConfirmClose should only be called when
closing a tab while the application is *not* shutting down. However
this isn't a real fix in the sense that the issue would be now
resolved for once and for all but in fact the same problem can appear
at any time some temporary Qt event loop is entered. (I.e. when
QMessageBox or modal QDialog executes). Easiest real fix would be
to *not* post IterateGameLoopEvents at all. But that'd have the
problem that background interactivity (such as rendering) of the
widgets when a dialog (such as color chooser) is open would be lost.
It's hard to think of a systematic fix for this category of bugs,
so these might need to be fixed case-by-case basis.