Refactor ChildWindow closing sequence
Refactor ChildWindow closing sequence. It seems that there's a
difference between calling QWidget::close and the user clicking the
X button of the window in terms of when the actual window object
(or the underlying X11 rendering surface) is destroyed. Try to
handle these now separately and make sure that in either case the
widget object (and the GfxWidget) is shutdown while the window
still exists and is valid.