Refactor saving of workspace resources to use class object ID
Refactor saving of workspace resources to use class object ID
instead of resource name. This simplifies the logic of saving
since there's no longer a need to check for name collision and
overwriting an existing resource object. This however also means
that there can now be multiple resource / class objects of the
same type and using the same name. This means that the resource
loading/resolution also needs to change to using IDs and also
that the UIs need to provide means to identify which resource
the user wants to use when there are multiple resources with the
same name. Both of these changes are left for later commits.