Refactor resource addressing to use class object IDs instead of names
Refactor resource addressing to use class object IDs instead of names.
The class object IDs are immutable and this change simplifies the
resource management in the sense that the mutable properties (such
as the resource name) can change independently without breaking the
existing resource links.

The UI will still present the user human readable names for resources
(i.e the names associated with class objects). As of now there doesn't
exist a method to define which resource to use in case there are
multiple resources with the same name. This needs to be addressed
in some later commit.