Refactor the main interface between game host and game library
Refactor the main interface between game host and game library.
This change removes the mention of base::Logger from the main
game entry point (MakeApp) and instead provides a new function
for the host to call to set the logger object. Also a new function
is added for setting the resource loader for the gfx subsystem.

API changes also reflected in the editor's play window.