Invaders: Refactor several game animations to use gamelib animations
Invaders:  Refactor several game animations to use gamelib animations
that are created based on the resources/data created in the editor.

This change unfortunately breaks some of the functionality related
to scaling of the content. Scaling content as the window size
changes needs to be rethought when all the game content has moved
on to using gamelib based animations.