Add process separation and IPC channels for game playback
Add process separation and IPC channels for game playback. When
a flag is set to enable separate process for the game playback
a new process is spawned. The EditorGameHost will proceed to load
the game library and invoke functions on the game application as
usual. The great benefit is that when the game crashes the editor
is not crashing with it but only the game host surrogate process
will be taken down by the OS. The downside is that any changes done
to the workspace content need now be reflected to the surrogate
game process through IPC messaging. For game resources the initial
state is loaded from the workspace content files and then subsequent
changes are sent over IPC.

Note that currently there's a bug that when launching the separate
process the playback window geometry restore doesn't work on Linux.
The window appears in random positions on the desktop. Calling
move on it doesn't work either (Qt itself seems to be confused about
the window position).

Also it's possible to launch EditorGameHost as a standalone executor
for the game. Lauch with --help flag to see a list of options.