Add editor style setting in the application settings
Add editor style setting in the application settings and remove
command line flag for the darkstyle. The darkstyle is now presented
as a one of the styles available in the settings dialog. Likewise
the command line parameter is removed from the game host and instead
the editor will pass the style argument on the command line when
invoking the surrogate process.