Fold material uniform parameters into constants in GLSL
Fold material uniform parameters into constants in GLSL when the
material is static. There's a number of parameters that can be
treated this way in order to improve efficiency, i.e. to reduce
the overhead of having to call glUniformX. However this means that
these material programs will now become separate programs consuming
more GPU memory. Using static flag for a material is therefore
a trade off between performance and resource consumption.

Current implementation is very silly and doesn't really understand
GLSL syntax properly. Instead it does a simple textual translation
of lines of text in the shader code when something interesting is
found in the text. This is a little bit silly but works for now.