Improve vertex insertion logic in polygon widget on mouse dbl click
Improve vertex insertion logic in polygonw idget on mouse dbl click.
Use the first vertex of the triangle fan along with the vertex
closes to the click point and the new vertex to determine the
winding order for the new triangle and then use the winding order
to determine whether to insert before or after the closest vertex.

Note that this still isn't 100% since it's possible to have a
degenerate case where the winding order cannot be determined. For
example when the vertex closest to the click point is the first
vertex of the triangle fan.

Remember more UI settings across load/save resource.