Fix text rasterization bitmap caching bug
Fix text rasterization bitmap caching bug. Add a missing check for
checking if the cached bitmap is currently in use or not. Since the
key to the bitmap cache is based on width/height it's possible that
a line that would rasterize to identically dimensioned bitmap could
pickup a a cached copy of a bitmap while that is held in the vector
holding all the rasterized lines before compositing them together.