Refactor draw functionality out of the animation system
Refactor draw functionality out of the animation system. This change
simplifies the animation system by removing the messy dependency
to runtime objects such as material and drawable instances and class
object instances. This then in turn removes a lot of the questions
such as what to do when copying or cloning an animation node or
a animation node class object.

Additionally this allows the rendering functionality to vary
independently of the animation system. I.e. there can now easily be
several different rendering implementations all producing a different
rendered output of the same animation hieararchy.

Additionally the assumption is that the renderer can be also be used
and leveraged to render the eventual scene graph that is coming soon.

Further work might involve splitting the animation track functionality
out of the animation class as well but that'd a later commit.