Add the first version of scene functionality
Add the first version of scene functionality. Scene is the top
level representation of a "game world". It contains nodes that are
placed in a hierarchical transformation tree (aka render tree,
aka scene graph). Each node can contain a number of items pertaining
to the node and its characteristics. An example of such item could
be a rigid body which defines the node in terms of it's physics
properties. A subsystem (in this case the physics engine) will then
look at these properties and act accordingly on the node and on the
scene possibily transforming the scene in some way. For example the
physics engine would reposition/reorient the scene nodes based on
the latest physics simulation data.

There's currently significant overlap in the functionality between
Scene and Animation (and SceneNode and AnimationNode) but it's
uncertain whether these systems will converge or diverge. Therefore
these are now left as is and are separate types. However in addition
to current functionality scene is likely to acquire more complicated
functionality for example building acceleration structures for quickly
determinining which objects are within some bounding volume (rectangle).