Relax the requirement for the item to be bound in UI combobox
Relax the requirement for the item to be found in UI combobox. In
some sense if the item that we're trying to select in the combox
isn't found in the combobox it is a bug. However it's difficult
to make sure that every no code path would always correctly replace
or migrate some setting from old value to new value. However the
combobox also allows -1 to be used (it goes into "no item selected")
state. In summary failing the assert would in most cases result
only in a minor UI bug so making the decision for now to relax this
condition and remove the assert.