Add first version of "default" game engine implementation
Add first version of "default" game engine implementation. Engine
will be the component that will orchestrate the functionality
between various subsystems such as physics simulation, rendering,
scripting and audio engine. This will be the *default* implementation
that is hopefully sufficient for most games. The actual game logic
will then be implemented in scripting (in Lua scripts) with bindings
to access the scene and some of the subsystems. The engine is built
into a shared library that implements the main interface between the
game host (game main) and the loadable game module.

Games however are of course welcome to implement their own engines.
Those engines will then need to be build using a C++ compiler. The
project settings can then be tuned to use a "Custom" engine and
include a path to the engine library that will be used.

Note that this is *very* rough first version of the engine but it's
already useful in this current bare bones state because it allows
more rapid development of features such as physics because it's now
possible to quickly edit a scene in the editor and then launch the
engine with a scene.