Transform coordinates from physics world into game world properly
Transform coordinates from physics world into game world properly
by transforming them into the node's coordinate frame instead of
assuming the glogal game world frame. I.e. if a node is a child
of another node its global physics transformation must be expressed
as a transformation relative to its parent instead of being relative
to the global transformation frame.

This also requires the proper traversal of nodes in the world in order
to make sure that the parent nodes are always transformed before
their children.

There's a whole bunch of complexity in this operation and also a
fair amount of matrix operations. It might make sense to have an
optimized path for objects that are in the global coordinate frame
in order to cut out some of the matrix operations.

There also seems to be some instability when dynamic bodies are in
a hiearchy. This can be seen in the rendered objects not being quite
where the physics object would. Not sure yet which causes this.