Refactor/rename gamelib scene classes into Entity classes
Refactor/rename gamelib scene classes into entity classes. Previous
plan was to have animated characters be represented in the scene as
scene node attachments but that has the problem that it's difficult
to create more sophisticated objecs with complex shapes for physics
simulation. This can be addressed by creating a new concept and
layer of abstraction called Entity. The entity class will hold an
hieararchy of nodes (the render tree) and some attachments (items)
for each node. Then the scene class will contain references to the
entity classes. When the scene's physics simulation and rendering
will take place all the entities and their render trees are then
traversed as part of the whole scene traversal in order to transform
them into the rendering/physics space. This allows more complicated
entites with more complicated structure and more sophisticated
physics handling for example for collision detection. I.e. entity's
nodes might have rigid bodies enabled with collision boxes. Thus
in the physics world the entity will in fact have multiple collision
sensors.

Currently the entity system has a lot of overlap with the animation
system and the plan is to refactor the remaining non-overlapping
functionality (the animation track system) from animation class
and merge it into the entity class and then remove the animation
class entirely.